Fix functionality for PHP 8.0
"For backward compatibility reasons, array_key_exists() will also return
true if key is a property defined within an object given as array. This
behaviour is deprecated as of PHP 7.4.0, and removed as of PHP 8.0.0."
- https://www.php.net/manual/en/function.array-key-exists

array_key_exists() was replaced with property_exists() as it is an
object being checked and not an array.